break renderCharacters logic up, and allow user to move a character to the arena
