+++ src/index.js
@@ -1,6 +1,6 @@
-import { createCharacters } from "./javascript/main.js";
+import { createCharacters, selectOneCharacter } from "./javascript/main.js";
 import "./css/reset.css";
 import "./css/style.css";
 
 createCharacters();
-console.log("works?");
+selectOneCharacter();
