+++ src/javascript/characterConstructor.js
