+++ src/javascript/domManipulation/selectCharacter.js
@@ -1,0 +1,1 @@
+export const selectCharacter = () => {};
