+++ src/javascript/main.js
@@ -1,6 +1,11 @@
-import { characters } from "./createCharacters";
+import { characters } from "./characterConstructor";
 import { renderCharacters } from "./render/renderCharacters";
+import { selectCharacter } from "./domManipulation/selectCharacter";
 
 export const createCharacters = () => {
   renderCharacters(characters);
+};
+
+export const selectOneCharacter = () => {
+  selectCharacter();
 };
