--- src/javascript/selectCharacter.js
@@ -1,4 +1,0 @@
-export const selectCharacter = () => {
-  let characters = document.getElementsByClassName("character");
-  console.log(characters);
-};
